[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:indexed_search
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Resolves: #92169
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65909
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/indexed_search/Classes/Domain/Repository/IndexSearchRepository.php
@@ -231,7 +231,7 @@ class IndexSearchRepository
             // Total search-result count
             $count = $result->rowCount();
             // The pointer is set to the result page that is currently being viewed
-            $pointer = MathUtility::forceIntegerInRange($this->resultpagePointer, 0, floor($count / $this->numberOfResults));
+            $pointer = MathUtility::forceIntegerInRange($this->resultpagePointer, 0, (int)floor($count / $this->numberOfResults));
             // Initialize result accumulation variables:
             $c = 0;
             // Result pointer: Counts up the position in the current search-result
